add connect wallet
